Move getSiteCommentsTree
